+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [0.1.8] (2026-01-01)
 ### Added
 * the `Bicmid` table into databse
 * `bicmid` on browser's local strage
@@ -17,7 +20,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * primary key AUTOINCREMENT on create table
 * get_or_store_xxx() must return -1 if it errors
-
 
 ## [0.1.7] (2025-12-27)
 ### Changed
@@ -64,7 +66,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * first commit
 
-[Unreleased]: https://github.com/aki-akaguma/browserinfocm/compare/v0.1.7..HEAD
+[Unreleased]: https://github.com/aki-akaguma/browserinfocm/compare/v0.1.8..HEAD
+[0.1.8]: https://github.com/aki-akaguma/browserinfocm/compare/v0.1.7..v0.1.8
 [0.1.7]: https://github.com/aki-akaguma/browserinfocm/compare/v0.1.6..v0.1.7
 [0.1.6]: https://github.com/aki-akaguma/browserinfocm/compare/v0.1.5..v0.1.6
 [0.1.5]: https://github.com/aki-akaguma/browserinfocm/compare/v0.1.4..v0.1.5